requirements.txt
aiogram==3.24.0
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
aiohttp==3.13.2
beautifulsoup4==4.12.3
lxml==5.3.0
python-dotenv==1.0.1
